feat: determine combined state as hub(1) does
The GitHub API does not specify how the state property is related to
statuses.state, so it's not clear what differences it may have.  Behave
like hub(1), assuming this is what users expect and that there is reason
for its current behavior.

One notable change:  If a ref has no status (e.g. because it hasn't been
pushed to GitHub), statuses is empty and state is "pending" in the API
response.  With this change, the combined state becomes "", which has a
different exit code and is written to stdout as "no status".

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>