refactor: bin -> root -> lib
The bin script should only handle gathering options from the user (i.e.
from the command line and environment variables).  Move the logic for
processing the response to the package root.  Move the function which is
currently there, which fetches the status from the GitHub API, into lib.

Hopefully this will make the module more useful for anyone who wants to
use it programmatically.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>