Switch from yargs to commander
The problems which previously motivated me to switch from commander to
yargs have been solved (by the addition of .exitOverride() and
.configureOutput()).  Commander has more flexibility than yargs for
option processing using .on('option'), which can be used for options
which are sensitive to ordering or have more complicated handling.  It
is also much simpler and has less exotic behavior that needs to be
disabled (see all the yargs options which were set).  Finally, it is
about 1/6 the total size.

Also, for this project specifically, yargs@17.1.0 broke positional
argument parsing due to yargs/yargs#1977.  Since the argument is
optional, `.demand()` is not appropriate (it's also deprecated).  It
appears the correct fix would be to add a default command and define the
positional argument on that.  However, I'm done dealing with yargs
breakage, and switching to Commander will a better return on effort.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>